+++ Packages/com.unity.cloud.gltfast.tests/Tests/Runtime/Scripts/Import/AssetsTests.cs
@@ -20,7 +20,7 @@ namespace GLTFast.Tests.Import
     [TestFixture, Category("Import")]
     class AssetsTests
     {
-        [GltfTestCase("glTF-test-models", 57)]
+        [GltfTestCase("glTF-test-models", 58)]
         public IEnumerator GltfTestModels(GltfTestCaseSet testCaseSet, GltfTestCase testCase)
         {
             yield return AsyncWrapper.WaitForTask(RunTestCase(testCaseSet, testCase));