+++ Packages/com.unity.cloud.gltfast/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- (Performance) Import mesh indices as unsigned integers and don't convert to signed integers anymore.
 
 ### Fixed